requirements.txt
pyrofork
quart
speedtest-cli
uvicorn
python-dotenv
psutil
asyncpg
tgcrypto